saleor/requirements.txt
# Saleor Backend Dependencies
Django>=4.2,<5.0
djangorestframework>=3.14.0
graphene-django>=3.0.0
django-filter>=23.0
celery>=5.3.0
redis>=5.0.0
psycopg2-binary>=2.9.0
Pillow>=10.0.0
django-cors-headers>=4.0.0
stripe>=7.0.0
babel>=2.12.0
pytz>=2023.3

# Testing
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
faker>=19.0.0

# Development
black>=23.0.0
flake8>=6.1.0
isort>=5.12.0

